+++ python/discopt/modeling/__init__.py
@@ -85,6 +85,7 @@ from discopt.modeling.core import (
 from discopt.modeling.core import (
     abs_ as abs,
 )
+from discopt.modeling.sets import ProductSet, RangeSet, Set
 
 __all__ = [
     "Model",
@@ -141,4 +142,7 @@ __all__ = [
     "atleast",
     "atmost",
     "exactly",
+    "Set",
+    "RangeSet",
+    "ProductSet",
 ]